requirements.txt
fastapi
uvicorn[standard]
redis
rq
Pillow
ultralytics
redis>=4.5.0
